fix(mesh): RTCM'in base'e girisi ve drone'a cikisi - 6 kusur
RTCM zinciri yazilmisti ama hic uctan uca calistirilmamisti. Calistirilinca
firmware tarafinda alti ayri kopukluk cikti; her biri sessizdi.

RX BASE - RTCM base'e hic giremiyordu:
- TIP_RTK dispatch whitelist'inde yoktu, cerceve "taninmayan tip" dalinda
  sessizce atiliyordu (hata sayaci bile artmiyordu)
- Ortak ayristirici tamponu 32B idi; RTCM cercevesi COBS ile ~1040B.
  Boyut artik derleme parametresi: RX BASE 1100, TX DRONE 32'de kaliyor
- uart_frame_parser_t.idx uint8_t idi -> 255 ustu cerceve sariyordu. Latent:
  ilk ikisi duzelmeden gorunmuyor, duzelince "buyuk cerceveler bozuk" diye
  RF'te aranirdi. Tasma kapisi sarma sonrasi hep dogru kaldigi icin hicbir
  sayaca yansimiyor
- loop() basina 32B okuma siniri: tek RTCM mesaji 32+ tur, CSMA gecikmeleriyle
  yuzlerce ms. YKI_LOOP_MAKS_BAYT=1200 (sinir kaldirilmadi, buyutuldu)

TX DRONE - RTCM ve failsafe yanlis porta yaziliyordu:
- rtk_mesh_loop() argumansiz cagriliyordu -> varsayilan Serial1. RPI_SERIAL0_MODU
  RPi hattini Serial0'a tasimisti; drone RTCM'i alip birlestirip RPi'ye BAGLI
  OLMAYAN porta yaziyordu
- ayni hata failsafe_kontrol()'de: mesh kopunca 0xFA RTL/LAND bildirimi RPi'ye
  hic ulasmiyordu, esp32_bridge::_isle_failsafe ateslenmiyordu. RX BASE ikisini
  de dogru yapiyordu, TX DRONE ikisini de atlamisti

Ayrica:
- mesh_log.h: paylasilan header'lardaki ~45 kosulsuz Serial cikisi, Serial'in
  veri hatti oldugu derlemelerde (RPI_SERIAL0_MODU/TEK_USB_MODU) binary akisi
  bozuyordu. MESH_LOG_* ile susturuldu; boot mesajlari bilerek birakildi
- RTCM fragmentleri artik canli node tablosuna UNICAST (802.11 ACK+retry).
  Broadcast'te tek parca kaybi TUM mesaji oldururdu. CSMA beklemesi fragment
  basina bir kez, hedef basina degil
- rtk_rtcm_payload_isle(): govde dogrulamasi tek yerde, iki giris yolu paylasiyor
- rtk_sender.h'deki bayat yorumlar duzeltildi (AES-GCM, 191B -> 238B)

Dogrulama: 32/32 birim testi (ASan+UBSan), eski uint8_t idx ile 8 hata verip
testin disli oldugu kanitlandi, 4/4 hedef env SUCCESS, binary'de string kontrolu
(serial0 env'i 2756B, tek_usb 3208B daha az flash - loglar gercekten dusmus).

+++ firmware/esp32_mesh/common/mesh_shared/uart_frame_parser.h
@@ -23,6 +23,15 @@
 // Pi<->ESP komut protokolu cerceveleri kucuktur (ham <= tip+id+18B payload+
 // crc16 = 22B; COBS + terminator <= ~27B). 32, eski main.cpp'lerdeki
 // rx_buf[32]/rpi_rx_buf[32] boyutuyla ayni ve yeterli marj birakiyor.
+//
+// AMA: RX BASE'in YKİ hatti coklanmis ve TIP_RTK de oradan geliyor. Bir RTCM3
+// mesaji en fazla 1029B (spec §2.6), cerceveye girince
+// TIP(1)+ID(1)+1029+CRC16(2) = 1033B, COBS worst-case ~1040B. 32 baytlik tampon
+// bunu alamaz; her RTCM cercevesi tasma dalinda (idx=0) sessizce dusuerdi.
+// Bu yuzden boyut derleme zamaninda buyutulebilir:
+//     RX BASE/platformio.ini  ->  -D UART_FRAME_BUF_SIZE=1100
+// TX DRONE'da 32 kaliyor: drone'un RPi hatti buyuk cerceve tasimiyor (RTCM
+// ters yonde, hava linkinden gelip _rtk_uart_gonder ile YAZILIYOR).
 #ifndef UART_FRAME_BUF_SIZE
 #define UART_FRAME_BUF_SIZE 32
 #endif
@@ -33,7 +42,11 @@ typedef struct {
     // decoded ve raw ayni boyutta; cobs_decode ciktisi giristen her zaman kisa
     // oldugundan tasma imkansiz.
     uint8_t decoded[UART_FRAME_BUF_SIZE];  // cozulmus cerceve (payload buraya isaret eder)
-    uint8_t idx;
+    // uint16_t: UART_FRAME_BUF_SIZE 255'i asabiliyor (bkz yukaridaki RTCM notu).
+    // uint8_t kalsaydi idx 256'da sarar, tasma kontrolu (idx < BUF_SIZE) hep
+    // dogru kalir ve ayristirici cerceveyi bastan yazmaya baslardi — semptom
+    // "buyuk cerceveler bozuk gelir", sebebi gorunmez.
+    uint16_t idx;
 } uart_frame_parser_t;
 
 static inline void uart_frame_parser_sifirla(uart_frame_parser_t* st) {
@@ -60,7 +73,7 @@ static inline bool uart_frame_parser_push(
         else st->idx = 0;   // tasma: cerceveyi at, yeniden senkronize ol
         return false;
     }
-    uint8_t n = st->idx;
+    uint16_t n = st->idx;
     st->idx = 0;            // 0x00'da her zaman sifirla (yapisal desync korumasi)
     if (n < 4) return false;
     uint16_t dlen = cobs_decode(st->raw, (uint16_t)n, st->decoded);
